Varie Modifiche Helper

+++ NuovoCinemaParadiso/Dtos/DtoBiglietto.cs
@@ -4,7 +4,7 @@ public class DtoBiglietto
     public string? Id { get; set; }
     public string ProiezioneId { get; set; } = string.Empty;
     public string UtenteId { get; set; } = string.Empty;
-    public decimal PrezzoFinale {get;set;}
+    public int PrezzoFinale {get;set;}
     public DateTimeOffset OrarioCreazione {get;set;} 
     public int NumeroBiglietti {get;set;}
     public string MetodoPagamento { get; set; } = "standard";
