+++ NuovoCinemaParadiso/Dtos/DtoCreazioneMovie.cs
@@ -13,7 +13,7 @@ public class DtoCreazioneMovie
     [Range(1, int.MaxValue, ErrorMessage = "La durata deve essere un numero intero positivo maggiore di 0.")]
     public int DurataMinuti { get; set; }
     [Range(0, double.MaxValue, ErrorMessage = "Il prezzo deve essere un numero positivo.")]
-    public decimal PrezzoMovie { get; set; }
+    public int PrezzoMovie { get; set; }
     [Required]
     public string GenereId { get; set; } = string.Empty;
 }
