+++ NuovoCinemaParadiso/Dtos/DtoMovie.cs
@@ -6,7 +6,7 @@ public class DtoMovie
     public string Titolo {get;set;} = string.Empty;
     public string Descrizione {get;set;} = string.Empty;
     public int DurataMinuti {get;set;}
-    public decimal PrezzoMovie {get;set;}
+    public int PrezzoMovie {get;set;}
     public string GenereId {get;set;} = string.Empty;
     public string Genere {get;set;} = string.Empty;
 }
