+++ NuovoCinemaParadiso/Dtos/DtoTipologiaSala.cs
@@ -4,5 +4,5 @@ public class DtoTipologiaSala
 {
     public string? Id { get; set; }
     public string Nome { get; set; } = string.Empty;
-    public decimal MaggiorazionePrezzo { get; set; }
+    public int MaggiorazionePrezzo { get; set; }
 }
